cluster layers interaction
