+++ libs/deck-layers/src/layers/cluster/index.ts
@@ -1,0 +1,2 @@
+export * from './ClusterLayer'
+export * from './cluster.types'
